requirements.txt
fastapi
pydantic-settings
uvicorn[standard]
chromadb
openai
pypdf